TODO/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==3.2.10
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
numpy==1.22.2
psycopg==3.0.9
psycopg2-binary==2.9.3
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
